+++ unlock-js/src/v6/index.js
@@ -8,6 +8,7 @@ import withdrawFromLock from './withdrawFromLock'
 import initializeTemplate from './initializeTemplate'
 import configureUnlock from './configureUnlock'
 import isLockManager from './isLockManager'
+import cancelAndRefund from './cancelAndRefund'
 
 export default {
   version: 'v6',
@@ -22,4 +23,5 @@ export default {
   withdrawFromLock,
   configureUnlock,
   isLockManager,
+  cancelAndRefund,
 }